refactor: TimeZoneClock seperated from AppHeader
